+++ examples/mini-beamline/src/point_detector/ioc_support.rs
@@ -70,6 +70,7 @@ impl DeviceSupport for PointDetectorDeviceSupport {
             let iface = match info.param_type {
                 RegistryParamType::Int32 => "asynInt32",
                 RegistryParamType::Float64 => "asynFloat64",
+                RegistryParamType::Int32Array => "asynInt32Array",
                 RegistryParamType::Float64Array => "asynFloat64Array",
                 RegistryParamType::OctetString => "asynOctet",
             };